requirements.txt
pyqt5==5.15.7
pyqt5-tools==5.15.4.3.2
pysnc==1.0.7
psycopg2==2.9.6
python-decouple==3.8
